Monstrous refactoring exercise
Mainly to split spritesheet up into snoo, logo, and everything else to keep the individual file sizes down. The image optimisation has been split into a lossy and lossless task - this is so the Snoo can be losslessly compressed (because lossy results in noticeable dithering), and everything else can be lossy. Also took the opportunity to rename a few background images that had long and silly names.